Make instruction component
